move jwtDecoder bean back to CelSecurityConfig
